requirements.txt
Flask==3.0.1
twilio==8.10.0
paramiko==3.4.0
selenium==4.15.2
python-dotenv==1.0.0
requests==2.31.0
pywhatkit==5.4
cryptography==41.0.7 